% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_methods.R
\name{coef.corrected_lasso}
\alias{coef.corrected_lasso}
\title{Extract Coefficients of a Corrected Lasso object}
\usage{
\method{coef}{corrected_lasso}(object, ...)
}
\arguments{
\item{object}{Fitted model object returned by \code{\link{corrected_lasso}}.}

\item{...}{Other arguments (not used).}
}
\description{
Default coef method for a \code{corrected_lasso} object.
}
