% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_methods.R
\name{coef.gds}
\alias{coef.gds}
\title{Extract Coefficients of a Generalized Dantzig Selector Object}
\usage{
\method{coef}{gds}(object, all = FALSE, ...)
}
\arguments{
\item{object}{Fitted model object returned by \code{\link{gds}}.}

\item{all}{Logical indicating whether to show all coefficient estimates, or only non-zeros.}

\item{...}{Other arguments (not used).}
}
\description{
Default coef method for a \code{gds} object.
}
