% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_methods.R
\name{coef.gmu_lasso}
\alias{coef.gmu_lasso}
\title{Extract Coefficients of a GMU Lasso object}
\usage{
\method{coef}{gmu_lasso}(object, all = FALSE, ...)
}
\arguments{
\item{object}{Fitted model object returned by \code{\link{gmu_lasso}}.}

\item{all}{Logical indicating whether to show all coefficient estimates, or
only non-zeros. Only used when delta is a single value.}

\item{...}{Other arguments (not used).}
}
\description{
Default coef method for a \code{gmu_lasso} object.
}
