% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_methods.R
\name{coef.gmus}
\alias{coef.gmus}
\title{Extract Coefficients of a GMUS object}
\usage{
\method{coef}{gmus}(object, all = FALSE, ...)
}
\arguments{
\item{object}{Fitted model object returned by \code{\link{gmus}}.}

\item{all}{Logical indicating whether to show all coefficient estimates, or
only non-zeros. Only used when delta is a single value.}

\item{...}{Other arguments (not used).}
}
\description{
Default coef method for a \code{gmus} object.
}
