% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.correctedLasso.R
\name{cv.correctedLasso}
\alias{cv.correctedLasso}
\title{Cross-validated Corrected lasso}
\usage{
cv.correctedLasso(W, y, sigmaUU, nfolds = 10, radius = NULL,
  noRadii = 100, alpha = 0.1, maxits = 5000, standardize = FALSE)
}
\arguments{
\item{W}{Design matrix, measured with error.}

\item{y}{Vector of the continuous response value.}

\item{sigmaUU}{Covariance matrix of the measurement error.}

\item{nfolds}{Number of folds to use in cross-validation. Default is 100.}

\item{radius}{Optional vector containing the set of radii of the l1-ball onto which the solution is projected.}

\item{alpha}{Optional step size of the projected gradient descent algorithm. Default is 0.1.}

\item{maxits}{Optional maxium number of iterations of the project gradient descent algorithm for each radius. Default is 5000.}
}
\value{
An object with S3 class "cv.correctedLasso", "list"
}
\description{
Cross-validated Corrected lasso
}
\details{
Corrected version of the lasso for the case of linear regression, estimated using cross-validation. The method does require an estimate of the measurement error covariance matrix.
}
\examples{
# Gaussian
set.seed(100)
n <- 100; p <- 50 # Problem dimensions
X <- matrix(rnorm(n * p), nrow = n)  # True (latent) variables
sigmaUU <- diag(x = 0.2, nrow = p, ncol = p) # Measurement error covariance matrix (typically estimated by replicate measurements)
W <- X + rnorm(n, sd = diag(sigmaUU)) # Measurement matrix (this is the one we observe)
beta <- c(seq(from = 0.1, to = 1, length.out = 5), rep(0, p-5)) # Coefficient
y <- X \%*\% beta + rnorm(n, sd = 1) # Response
cvfit <- cv.correctedLasso(W, y, sigmaUU) # Run the corrected lasso
plot(cvfit)
# Run the standard lasso using the radius found by cross-validation
fit <- correctedLasso(W, y, sigmaUU, family = "gaussian", radius = cvfit$radius.min)
}
\references{
Po-Ling Loh, and Martin J. Wainwright. 2012. "High-Dimensional Regression with Noisy and Missing Data: Provable Guarantees with Nonconvexity." The Annals of Statistics 40 (3) https://projecteuclid.org/euclid.aos/1346850068

Oystein Sorensen, Arnoldo Frigessi, and Magne Thoresen. 2015. "Measurement Error in Lasso: Impact and Likelihood Bias Correction." Statistica Sinica 25 (2).
}

