% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gmu_lasso.R
\name{fit_gmu_lasso}
\alias{fit_gmu_lasso}
\title{Generalized Matrix Uncertainty Lasso}
\usage{
fit_gmu_lasso(W, y, lambda, delta, family = "binomial", active_set = FALSE)
}
\arguments{
\item{W}{Design matrix, measured with error. Must be a numeric matrix.}

\item{y}{Vector of responses.}

\item{lambda}{Regularization parameter.}

\item{delta}{Additional regularization parameter, bounding the measurement
error.}

\item{family}{Character string. Currently "binomial" and "poisson" are supported.}

\item{active_set}{Logical. Whether or not to use an active set strategy to
speed up coordinate descent algorithm.}
}
\value{
Coefficient vector.
}
\description{
Generalized Matrix Uncertainty Lasso
}
\references{
\insertRef{rosenbaum2010}{hdme}

\insertRef{sorensen2018}{hdme}
}
