% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gaussian.R
\docType{data}
\name{gaussian}
\alias{gaussian}
\title{A simulated gaussian data set}
\format{A data frame with 300 rows and 8 columns.}
\usage{
gaussian
}
\description{
A simulated gaussian data set
}
\keyword{datasets}
