% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-logistic.R
\docType{data}
\name{logistic}
\alias{logistic}
\title{A simulated logistic data set}
\format{A data frame with 300 rows and 8 columns.}
\usage{
logistic
}
\description{
A simulated logistic data set
}
\keyword{datasets}
