% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musalgorithm.R
\name{musalgorithm}
\alias{musalgorithm}
\title{Algorithm for mus}
\usage{
musalgorithm(W, y, lambda, delta)
}
\arguments{
\item{W}{Matrix of measurements.}

\item{y}{Response vector.}

\item{lambda}{Regularization parameter due to residual.}

\item{delta}{Regularization parameter due to measurement error.}
}
\description{
Algorithm for mus
}
