% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musbinomial.R
\name{musbinomial}
\alias{musbinomial}
\title{Generalized Matrix Uncertainty Selector for logistic regression}
\usage{
musbinomial(W, y, lambda, delta)
}
\arguments{
\item{W}{Design matrix, measured with error.}

\item{y}{Vector of the binomial response value.}

\item{lambda}{Regularization parameter due to model error.}

\item{delta}{Regularization parameter due to measurement error.}
}
\value{
Intercept and coefficients at the values of lambda and delta specified.
 @import stats
}
\description{
Internal function.
}
