% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muselector.R
\name{muselector}
\alias{muselector}
\title{Generalized Dantzig Selector}
\usage{
muselector(W, y, lambda = NULL, delta = NULL, family = c("gaussian",
  "binomial"))
}
\arguments{
\item{y}{Vector of the continuous response value.}

\item{lambda}{Regularization parameter.}

\item{X}{Design matrix.}
}
\value{
Intercept and coefficients at the values of lambda and delta specified.
}
\description{
Generalized Dantzig Selector
}
\examples{
set.seed(1)
n <- 1000; p <- 10 # Problem dimensions
X <- matrix(rnorm(n * p), nrow = n) # True (latent) variables
W <- X + matrix(rnorm(n*p, sd = 1), nrow = n, ncol = p) # Measurement matrix (this is the one we observe)
beta <- c(seq(from = 0.1, to = 1, length.out = 5), rep(0, p-5))
y <- X \%*\% beta + rnorm(n, sd = 1) # Response
fit <- muselector(W, y) # Run the MU Selector
plot(fit) # Draw an elbow plot to select delta

# Now, according to the "elbow rule", choose the final delta where the curve has an "elbow".
# In this case, the elbow is at about delta = 0.12, so we use this to compute the final estimate:
fit <- muselector(W, y, delta = 0.12)
plot(fit) # Plot the coefficients

}
\references{
Emmanuel Candes and Terence Tao. 2007. "The Dantzig Selector: Statistical Estimation When p Is Much Larger Than n." The Annals of Statistics 35 (6) https://projecteuclid.org/euclid.aos/1201012958
}
