% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.corrected_lasso.R
\name{plot.corrected_lasso}
\alias{plot.corrected_lasso}
\title{plot.corrected_lasso}
\usage{
\method{plot}{corrected_lasso}(x, type = "nonzero")
}
\arguments{
\item{x}{Object of class corrected_lasso, returned from calling fit_corrected_lasso()}

\item{type}{Type of plot. Either "nonzero" or "path".}
}
\examples{
# Example with linear regression
n <- 100 # Number of samples
p <- 50 # Number of covariates
X <- matrix(rnorm(n * p), nrow = n)  # True (latent) variables
sigmaUU <- diag(x = 0.2, nrow = p, ncol = p) # Measurement error covariance matrix (typically estimated by replicate measurements)
W <- X + rnorm(n, sd = diag(sigmaUU)) # Measurement matrix (this is the one we observe)
beta <- c(seq(from = 0.1, to = 1, length.out = 5), rep(0, p-5)) # Coefficient
y <- X \%*\% beta + rnorm(n, sd = 1) # Response
fit <- fit_corrected_lasso(W, y, sigmaUU, family = "gaussian") # Run the corrected lasso
plot(fit)

}
