% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.correctedLasso.R
\name{plot.correctedLasso}
\alias{plot.correctedLasso}
\title{plot.correctedLasso}
\usage{
\method{plot}{correctedLasso}(x)
}
\arguments{
\item{x}{numeric number}

\item{...}{other arguments}
}
\examples{
a <- 1
class(a) <- "lm"
MyHappyFunction(a)
}

