% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cv_gds.R
\name{plot.cv_gds}
\alias{plot.cv_gds}
\title{plot.cv_gds}
\usage{
\method{plot}{cv_gds}(x, ...)
}
\arguments{
\item{x}{The object to be plotted, returned from \code{\link{cv_gds}}.}

\item{...}{Other arguments to plot (not used).}
}
\description{
Plot the output of \code{\link{cv_gds}}.
}
