% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gmus.R
\name{plot.gmus}
\alias{plot.gmus}
\title{Plot the estimates returned by fit_gmus and fit_mus}
\usage{
\method{plot}{gmus}(x)
}
\arguments{
\item{x}{An object of class gmus}
}
\description{
Plot the number of nonzero coefficients along a range of delta
  values if delta has length larger than 1, or the estimated coefficients of
  delta has length 1.
}
\examples{
# Example with linear regression
set.seed(1)
n <- 1000 # Number of samples
p <- 200 # Number of covariates
X <- matrix(rnorm(n * p), nrow = n) # True (latent) variables
W <- X + matrix(rnorm(n*p, sd = 0.4), nrow = n, ncol = p) # Measurement matrix (this is the one we observe)
beta <- c(seq(from = 0.1, to = 1, length.out = 5), rep(0, p-5)) # Coefficient vector
y <- X \%*\% beta + rnorm(n, sd = 1) # Response
mus1 <- fit_mus(W, y) # Run the MU Selector
plot(mus1) # Draw an elbow plot to select delta

# Now, according to the "elbow rule", choose the final delta where the curve has an "elbow".
# In this case, the elbow is at about delta = 0.08, so we use this to compute the final estimate:
mus2 <- fit_mus(W, y, delta = 0.08)
plot(mus2) # Plot the coefficients

}
