% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoissonRMoE.R
\name{PoissonRMoE}
\alias{PoissonRMoE}
\title{Penalized MLE for the Poisson regularized Mixture of Experts.}
\usage{
PoissonRMoE(Xmat, Ymat, K, Lambda, Gamma, option)
}
\arguments{
\item{Xmat}{The matrix data for the input.}

\item{Ymat}{Vector of the response variable.}

\item{K}{Number of expert classes.}

\item{Lambda}{Penalty value for the expert part.}

\item{Gamma}{Penalty value for the gating network.}

\item{option}{\code{option = 1}: using proximal Newton-type method;
\code{option = 0}: using proximal Newton method.}
}
\value{
A list with the following objects.
\itemize{
\item \code{PPara.txt} contains 2K-1 vectors: the first K vectors are vectors
of the experts, the remain are vectors of the gating network.
\item \code{PLOG.txt} the penalized log-likelihood value.
\item \code{PBIC.txt} the value of BIC.
\item \code{PMAXP.txt} the gating network's values for each observation.
\item \code{GSigma.txt} the value of sigma.
\item \code{PRestore data.txt} contains the input data and the classification
class (the last column) for each observation.
}
}
\description{
This function provides a penalized MLE for the Poisson regularized Mixture
of Experts (MoE) model corresponding with the penalty parameters Lambda,
Gamma.
}
