% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.corrected_lasso}
\alias{print.corrected_lasso}
\title{Print a Corrected Lasso object}
\usage{
\method{print}{corrected_lasso}(x, ...)
}
\arguments{
\item{x}{Fitted model object returned by \code{\link{corrected_lasso}}.}

\item{...}{Other arguments (not used).}
}
\description{
Default print method for a \code{corrected_lasso} object.
}
