% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.cv_gds}
\alias{print.cv_gds}
\title{Print a Cross-Validated GDS Object}
\usage{
\method{print}{cv_gds}(x, ...)
}
\arguments{
\item{x}{Fitted model object returned by \code{\link{cv_gds}}.}

\item{...}{Other arguments (not used).}
}
\description{
Default print method for a \code{cv_gds} object.
}
