% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.gmu_lasso}
\alias{print.gmu_lasso}
\title{Print a GMU Lasso object}
\usage{
\method{print}{gmu_lasso}(x, ...)
}
\arguments{
\item{x}{Fitted model object returned by \code{\link{gmu_lasso}}.}

\item{...}{Other arguments (not used).}
}
\description{
Default print method for a \code{gmu_lasso} object.
}
