% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.gmus}
\alias{print.gmus}
\title{Print a GMUS object}
\usage{
\method{print}{gmus}(x, ...)
}
\arguments{
\item{x}{Fitted model object returned by \code{\link{gmus}}.}

\item{...}{Other arguments (not used).}
}
\description{
Default print method for a \code{gmus} object.
}
