% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMoE-package.R
\docType{package}
\name{RMoE-package}
\alias{RMoE}
\alias{RMoE-package}
\title{RMoE: LASSO Regularized Mixture of Experts Models}
\description{
\code{RMoE} is a package containing regularized Mixture of Experts
models using the Lasso penalty.

\code{RMoE} contains the following Regularized Mixture-of-Experts models:

\itemize{
\item GaussianRMoE: Gaussian Regularixed Mixture of Experts;
\item LogisticRMoE: Logistic Regularixed Mixture of Experts;
\item PoissonRMoE: Poisson Regularixed Mixture of Experts.
}

To learn more about \code{RMoE}, start with the vignettes:
\code{browseVignettes(package = "RMoE")}
}
\references{
Huynh B. T., Chamroukhi F. 2019. \emph{Estimation and Feature Selection in Mixtures of Generalized Linear Experts Models.} \url{https://arxiv.org/abs/1907.06994}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fchamroukhi/HDME}
  \item Report bugs at \url{https://github.com/fchamroukhi/HDME/issues}
}

}
\author{
\strong{Maintainer}: Bao-Tuyen Huynh \email{baotuyen.dlu@gmail.com}

Authors:
\itemize{
  \item Faicel Chamroukhi \email{faicel.chamroukhi@unicaen.fr} (0000-0002-5894-3103)
}

}
