% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cleveland.R
\docType{data}
\name{cleveland}
\alias{cleveland}
\title{The Cleveland data set}
\format{A data frame with 297 rows and 15 columns.}
\source{
\href{https://archive.ics.uci.edu/ml/datasets/heart+Disease}{https://archive.ics.uci.edu/ml/datasets/heart+Disease}
}
\usage{
cleveland
}
\description{
The Cleveland data set described on the website UC Irvine Machine Learning
Repository.
}
\keyword{datasets}
