% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gmu_lasso.R
\name{fit_gmu_lasso}
\alias{fit_gmu_lasso}
\title{Generalized Matrix Uncertainty Lasso}
\usage{
fit_gmu_lasso(W, y, lambda = NULL, delta = NULL, family = "binomial",
  active_set = TRUE)
}
\arguments{
\item{W}{Design matrix, measured with error. Must be a numeric matrix.}

\item{y}{Vector of responses.}

\item{lambda}{Regularization parameter.}

\item{delta}{Additional regularization parameter, bounding the measurement
error.}

\item{family}{Character string. Currently "binomial" and "poisson" are supported.}

\item{active_set}{Logical. Whether or not to use an active set strategy to
speed up coordinate descent algorithm.}
}
\value{
Coefficient vector.
}
\description{
Generalized Matrix Uncertainty Lasso
}
\examples{
n <- 200
p <- 50
s <- 10
beta <- c(rep(1,s),rep(0,p-s))
sdU <- 0.2

X <- matrix(rnorm(n*p),nrow = n,ncol = p)
W <- X + sdU * matrix(rnorm(n * p), nrow = n, ncol = p)

y <- rbinom(n, 1, (1 + exp(-X\%*\%beta))**(-1))
bNew <- fit_gmu_lasso(W, y)


}
\references{
\insertRef{rosenbaum2010}{hdme}

\insertRef{sorensen2018}{hdme}
}
