% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussRMoE.R
\name{GaussRMoE}
\alias{GaussRMoE}
\title{Penalized MLE for the regularized Mixture of Experts.}
\usage{
GaussRMoE(Xm, Ym, K, Lambda, Gamma, option = FALSE, verbose = FALSE)
}
\arguments{
\item{Xm}{The matrix data for the input.}

\item{Ym}{Vector of the response variable.}

\item{K}{Number of expert classes.}

\item{Lambda}{Penalty value for the expert part.}

\item{Gamma}{Penalty value for the gating network.}

\item{option}{Optional. \code{option = TRUE}: using proximal Newton-type method;
\code{option = FALSE}: using proximal Newton method.}

\item{verbose}{Optional. A logical value indicating whether or not values of
the log-likelihood should be printed during EM iterations.}
}
\value{
GaussRMoE returns an object of class \link{GRMoE}.
}
\description{
This function provides a penalized MLE for the regularized Mixture of Experts
(MoE) model corresponding with the penalty parameters Lambda, Gamma.
}
\seealso{
\link{GRMoE}
}
