% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muselector.R
\name{muselector}
\alias{muselector}
\title{Matrix Uncertainty Selector (MU Selector)}
\usage{
muselector(W, y, lambda = NULL, delta = NULL, family = c("gaussian",
  "binomial"))
}
\arguments{
\item{W}{Design matrix, measured with error.}

\item{y}{Vector of the continuous response value.}

\item{lambda}{Regularization parameter due to model error.}

\item{delta}{Regularization parameter due to measurement error.}
}
\value{
Intercept and coefficients at the values of lambda and delta specified.
}
\description{
Matrix Uncertainty Selector (MU Selector) for linear regression, a modified version of the Dantzig Selector for the case of covariate measurement error. The method was first proposed by Rosenbaum and Tsybakov (2010). The MU Selector does not require an estimate of the measurement error covariance matrix, but instead takes measurement error into account through an additional regularization parameter, delta.
}
\examples{
set.seed(1)
n <- 100; p <- 50 # Problem dimensions
X <- matrix(rnorm(n * p), nrow = n) # True (latent) variables
W <- X + matrix(rnorm(n*p, sd = 1), nrow = n, ncol = p) # Measurement matrix (this is the one we observe)
beta <- c(seq(from = 0.1, to = 1, length.out = 5), rep(0, p-5))
y <- X \%*\% beta + rnorm(n, sd = 1) # Response
fit <- muselector(W, y) # Run the MU Selector
plot(fit) # Draw an elbow plot to select delta

# Now, according to the "elbow rule", choose the final delta where the curve has an "elbow".
# In this case, the elbow is at about delta = 0.12, so we use this to compute the final estimate:
fit <- muselector(W, y, delta = 0.12)
plot(fit) # Plot the coefficients

}
\references{
Emmanuel Candes and Terence Tao. 2007. "The Dantzig Selector: Statistical Estimation When p Is Much Larger Than n." The Annals of Statistics 35 (6) https://projecteuclid.org/euclid.aos/1201012958

Mathieu Rosenbaum and Alexandre B. Tsybakov. 2010. "Sparse Recovery Under Matrix Uncertainty." The Annals of Statistics 38 (5) https://projecteuclid.org/euclid.aos/1278861455
}

