% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-poisson.R
\docType{data}
\name{poisson}
\alias{poisson}
\title{A simulated poisson data set}
\format{A data frame with 300 rows and 8 columns.}
\usage{
poisson
}
\description{
A simulated poisson data set
}
\keyword{datasets}
