% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoissonRMoE.R
\name{PoissonRMoE}
\alias{PoissonRMoE}
\title{Penalized MLE for the Poisson regularized Mixture of Experts.}
\usage{
PoissonRMoE(Xmat, Ymat, K, Lambda, Gamma, option = FALSE,
  verbose = TRUE)
}
\arguments{
\item{Xmat}{The matrix data for the input.}

\item{Ymat}{Vector of the response variable.}

\item{K}{Number of expert classes.}

\item{Lambda}{Penalty value for the expert part.}

\item{Gamma}{Penalty value for the gating network.}

\item{option}{Optional. \code{option = TRUE}: using proximal Newton-type method;
\code{option = FALSE}: using proximal Newton method.}

\item{verbose}{Optional. A logical value indicating whether or not values of
the log-likelihood should be printed during EM iterations.}
}
\value{
PoissonRMoE returns an object of class \link{PRMoE}.
}
\description{
This function provides a penalized MLE for the Poisson regularized Mixture
of Experts (MoE) model corresponding with the penalty parameters Lambda,
Gamma.
}
\seealso{
\link{PRMoE}
}
