% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-residential.R
\docType{data}
\name{residential}
\alias{residential}
\title{The Residential Building data set}
\format{A data frame with 372 rows and 109 columns.}
\usage{
residential
}
\description{
The Residential Building data set
}
\keyword{datasets}
